test: split the args.test.ts aggregation along source topology
AGENTS.md file-size tripwires now apply to tests with no exemption, and
test files are expected to mirror source topology 1:1. args.test.ts was
a 2,503-line aggregation in src/utils/__tests__ while the code it
exercises lives in src/cli/parser. Split it into six focused files with
every test moved verbatim (142 tests before and after):

- src/cli/parser/__tests__/args-parse-interaction.test.ts (29 tests):
  parseArgs shapes for press/click/swipe/gesture/type/record/screenshot
  and friends
- src/cli/parser/__tests__/args-parse-session.test.ts (41 tests):
  parseArgs shapes for session/daemon/device flags, passthrough,
  install/metro/connect/proxy/auth and friends
- src/cli/parser/__tests__/args-validation.test.ts (17 tests): strict/
  compat modes, rejections, deterministic errors
- src/cli/parser/__tests__/cli-help-topics.test.ts (15 tests): global
  usage and help topics
- src/cli/parser/__tests__/cli-help-command-usage.test.ts (35 tests):
  per-command usage copy
- src/utils/__tests__/command-schema-guards.test.ts (5 tests): schema/
  catalog/capability guards and the cli.ts dispatch-literal walk (the
  oxc-parser helpers live here)

AGENTS.md testing-matrix and help-source pointers updated to the new
paths, including the stale src/utils/cli-help.ts and cli-flags.ts
locations (both live under src/cli/parser/).
