+++ src/cli/parser/__tests__/args-validation.test.ts
@@ -1,0 +1,176 @@
+import { test } from 'vitest';
+import assert from 'node:assert/strict';
+import { parseArgs } from '../args.ts';
+import { AppError } from '../../../kernel/errors.ts';
+
+test('parseArgs rejects test retries above the supported ceiling', () => {
+  assert.throws(
+    () => parseArgs(['test', './suite', '--retries', '4'], { strictFlags: true }),
+    (error: unknown) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      /Invalid retries: 4/.test(error.message),
+  );
+});
+
+test('parseArgs rejects --launch-args on commands that do not allow it', () => {
+  assert.throws(
+    () => parseArgs(['tap', '100', '200', '--launch-args', 'foo'], { strictFlags: true }),
+    (error) => error instanceof AppError && error.code === 'INVALID_ARGS',
+  );
+});
+
+test('parseArgs rejects invalid record --fps range', () => {
+  assert.throws(
+    () => parseArgs(['record', 'start', './capture.mp4', '--fps', '0'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message === 'Invalid fps: 0',
+  );
+});
+
+test('parseArgs rejects invalid swipe pattern', () => {
+  assert.throws(
+    () => parseArgs(['swipe', '0', '0', '10', '10', '--pattern', 'diagonal']),
+    /Invalid pattern/,
+  );
+});
+
+test('parseArgs rejects conflicting back mode flags', () => {
+  assert.throws(
+    () => parseArgs(['back', '--in-app', '--system'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message ===
+        'back accepts only one explicit mode flag: use either --in-app or --system.',
+  );
+});
+
+test('debug rejects unrelated diagnostics flags', () => {
+  assert.throws(
+    () => parseArgs(['debug', 'symbols', '--include', 'headers']),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message.includes('not supported for command debug'),
+  );
+});
+
+test('compat mode warns and strips unsupported command flags', () => {
+  const parsed = parseArgs(['press', '10', '20', '--pause-ms', '2'], { strictFlags: false });
+  assert.equal(parsed.command, 'press');
+  assert.equal(parsed.flags.pauseMs, undefined);
+  assert.equal(parsed.warnings.length, 1);
+  assert.match(parsed.warnings[0]!, /not supported for command press/);
+});
+
+test('strict mode rejects unsupported pilot-command flags', () => {
+  assert.throws(
+    () => parseArgs(['press', '10', '20', '--pause-ms', '2'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message.includes('not supported for command press'),
+  );
+});
+
+test('strict mode rejects Metro override flags on doctor', () => {
+  assert.throws(
+    () => parseArgs(['doctor', '--metro-port', '9090'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message.includes('not supported for command doctor'),
+  );
+});
+
+test('strict mode rejects removed secondary alias', () => {
+  assert.throws(
+    () => parseArgs(['click', '@e5', '--secondary'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message === 'Unknown flag: --secondary',
+  );
+});
+
+test('strict mode rejects click-only button flag on press', () => {
+  assert.throws(
+    () => parseArgs(['press', '10', '20', '--button', 'secondary'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message.includes('not supported for command press'),
+  );
+});
+
+test('unknown short flags are rejected', () => {
+  assert.throws(
+    () => parseArgs(['press', '10', '20', '-x'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message === 'Unknown flag: -x',
+  );
+});
+
+test('negative numeric positionals are accepted without -- separator', () => {
+  const typed = parseArgs(['type', '-123'], { strictFlags: true });
+  assert.equal(typed.command, 'type');
+  assert.deepEqual(typed.positionals, ['-123']);
+
+  const typedMulti = parseArgs(['type', '-123', '-456'], { strictFlags: true });
+  assert.equal(typedMulti.command, 'type');
+  assert.deepEqual(typedMulti.positionals, ['-123', '-456']);
+
+  const pressed = parseArgs(['press', '-10', '20'], { strictFlags: true });
+  assert.equal(pressed.command, 'press');
+  assert.deepEqual(pressed.positionals, ['-10', '20']);
+});
+
+test('command-specific flags without command fail in strict mode', () => {
+  assert.throws(
+    () => parseArgs(['--depth', '3'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message.includes('requires a command that supports it'),
+  );
+});
+
+test('command-specific flags without command warn and strip in compat mode', () => {
+  const parsed = parseArgs(['--depth', '3'], { strictFlags: false });
+  assert.equal(parsed.command, null);
+  assert.equal(parsed.flags.snapshotDepth, undefined);
+  assert.equal(parsed.warnings.length, 1);
+  assert.match(parsed.warnings[0]!, /requires a command that supports/);
+});
+
+test('all commands participate in strict command-flag validation', () => {
+  assert.throws(
+    () => parseArgs(['open', 'Settings', '--depth', '1'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message.includes('not supported for command open'),
+  );
+});
+
+test('invalid range errors are deterministic', () => {
+  assert.throws(
+    () => parseArgs(['snapshot', '--backend', 'xctest'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message === 'Unknown flag: --backend',
+  );
+  assert.throws(
+    () => parseArgs(['snapshot', '--depth', '-1'], { strictFlags: true }),
+    (error) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      error.message === 'Invalid depth: -1',
+  );
+});
